feat: 다이어그램 뷰의 투두 블럭에도 원형 표시 대신 체크박스 사용. 테이블 뷰와 같은 인터렉션 추가
